Move vehicle inspection data to its own directory
