+++ examples/vehicle-inspection/src/vehicle-inspection/VehicleInspectionFormResult.ts
@@ -1,0 +1,7 @@
+import {Result} from '@raincatcher/wfm';
+import {VehicleInspectionFormData} from './VehicleInspectionFormData';
+export class VehicleInspectionFormResult implements Result {
+  public type = 'VehicleInspectionFormResult';
+  constructor(public data: VehicleInspectionFormData) {
+  }
+}
